Modification of the Discrete Clock adapter. Added some regression tests.

+++ ptolemy/cg/adapter/generic/program/procedural/c/adapters/ptolemy/actor/lib/SingleEvent.c
@@ -1,5 +1,5 @@
 /***initBlock($intPart, $fracPart)***/
-$fireAt(&director, $actorName(), $intPart + $fracPart/1000000000, 0);
+$fireAt(&director, $actorName(), $intPart + $fracPart/(double)1000000000, 0);
 /**/
 
 /***fireBlock($value)***/
